refactor: Move handlers for manager into manager folder
