+++ cmd/slides-to-video-manager/handlers/decorators.go
