+++ cmd/slides-to-video-manager/handlers/image_handlers.go
